refactored application startup, discord handlers and endpoint handlers
