+++ internal/endpoints.go
@@ -76,14 +76,11 @@ func StartAPI(db *bun.DB, dg *discordgo.Session) {
 	router.POST("/playdate/:id/maybe", api.setPlayDateAttendence)
 	router.POST("/playdate/:id/no", api.setPlayDateAttendence)
 
-	// Start discord handlers
-	dg.AddHandler(func(s *discordgo.Session, r *discordgo.MessageReactionAdd) {
-		api.setPlayDateAttendenceFromDisc(r.MessageReaction)
-	})
-	api.sendPatchNotes()
-
 	go api.watchDog()
-	router.Run("0.0.0.0:8080")
+	err := router.Run("0.0.0.0:8080")
+	if err != nil {
+		log.Err(err).Msg("failed to start web server")
+	}
 }
 
 type BaseTemplateData struct {
@@ -111,10 +108,9 @@ func (a *Api) watchDog() {
 
 	go func() {
 		for {
-			select {
-			case <-ticker.C:
-				a.fetchPoppedDates()
-			}
+			// fancy syntax to avoid switch case: https://github.com/dominikh/go-tools/issues/503
+			<-ticker.C
+			a.fetchPoppedDates()
 		}
 	}()
 }
@@ -343,96 +339,6 @@ func (a *Api) setPlayDateAttendence(c *gin.Context) {
 	state["PlayDatePlayers"] = playdatePlayers
 	state["PlayDate"] = playdate
 	c.HTML(http.StatusOK, "partials/players-table.html", state)
-}
-
-func (a *Api) setPlayDateAttendenceFromDisc(r *discordgo.MessageReaction) {
-	if r.UserID == "1252426978313633812" {
-		log.Debug().Msg("Reaction created by bot")
-		return
-	}
-
-	log.Info().Any("Reaction", r).Msg("Setting player attendance")
-	discId := r.UserID
-	react := r.Emoji.Name
-	msg, err := a.dg.ChannelMessage(Config.DiscordConfig.ChannelID, r.MessageID)
-	if err != nil {
-		log.Err(err).Msg("failed to get reaction message")
-		return
-	}
-	if msg.Author.ID != "1252426978313633812" {
-		log.Debug().Msg("Not a bot message")
-		return
-	}
-
-	attendance := AttendanceFrom(react) // parse input attendence action to internal enum
-	msgSplit := strings.Split(msg.Content, "/")
-	if len(msgSplit) <= 1 {
-		log.Debug().Msg("Not a playdate")
-		return
-	}
-	pId, err := strconv.Atoi(msgSplit[len(msgSplit)-1])
-	if err != nil {
-		log.Err(err).Msg("failed to parse given playdate id")
-		return
-	}
-
-	playdate := &PlayDate{ID: pId}
-	err = a.db.NewSelect().Model(playdate).WherePK().Scan(a.ctx)
-	if err != nil {
-		log.Err(err).Int("playdateID", pId).Msg("failed to find playdate")
-		return
-	}
-	if playdate.Status != PlayDateStatusPending {
-		log.Debug().Msg("PlayDate already happened")
-		return
-	}
-
-	log.Info().Int("playdateID", playdate.ID).Str("discordId", discId).Any("action", attendance).Msg("attempting to set playdate attendance")
-	player := &Player{DiscordID: discId}
-	err = a.db.NewSelect().Model(player).Where("discord_id = ?", player.DiscordID).Scan(a.ctx)
-	if err != nil {
-		log.Err(err).Str("discID", discId).Msg("failed to find player")
-		a.dg.ChannelMessageSend(Config.DiscordConfig.ChannelID, "Please go here to make an account: https://playdate.colinthatcher.dev/discord/login")
-		err = a.dg.MessageReactionRemove(Config.DiscordConfig.ChannelID, msg.ID, r.Emoji.APIName(), discId)
-		if err != nil {
-			log.Err(err).Msg("Failed to remove reaction on anon user")
-		}
-		return
-	}
-	rel := &PlayDateToPlayer{PlayDateID: playdate.ID, PlayerID: player.ID, Attending: attendance}
-	_, err = a.db.NewInsert().Model(rel).On("CONFLICT (playdate_id, player_id) DO UPDATE").Set("attending = EXCLUDED.attending").Exec(a.ctx)
-	if err != nil {
-		// send error back to user within the players-table.html
-		log.Error().Err(err).Interface("relation", rel).Msg("failed to insert playdate to player relation")
-	} else {
-		log.Info().Interface("relation", rel).Msg("successfully inserted playdate to player relation")
-	}
-
-	playdatePlayers := []*PlayDateToPlayer{}
-	err = a.db.NewSelect().Model(&playdatePlayers).Relation("Player").Where("playdate_id = ?", playdate.ID).Scan(a.ctx)
-	if err != nil {
-		// report error back to user, but just render the page like normal
-		log.Err(err).Any("playdate", playdate).Msg("failed to find related players to playdate")
-	}
-
-	if r.Emoji.APIName() != "👍" {
-		err = a.dg.MessageReactionRemove(Config.DiscordConfig.ChannelID, msg.ID, "👍", discId)
-		if err != nil {
-			log.Err(err).Str("Reaction", "👍").Msg("Failed to remove reaction")
-		}
-	}
-	if r.Emoji.APIName() != "🤔" {
-		err = a.dg.MessageReactionRemove(Config.DiscordConfig.ChannelID, msg.ID, "🤔", discId)
-		if err != nil {
-			log.Err(err).Str("Reaction", "🤔").Msg("Failed to remove reaction")
-		}
-	}
-	if r.Emoji.APIName() != "👎" {
-		err = a.dg.MessageReactionRemove(Config.DiscordConfig.ChannelID, msg.ID, "👎", discId)
-		if err != nil {
-			log.Err(err).Str("Reaction", "👎").Msg("Failed to remove reaction")
-		}
-	}
 }
 
 func (a *Api) registerUserTemplate(c *gin.Context) {
@@ -780,28 +686,6 @@ func getGithubReleaseNotes() (g GitHubRelease) {
 	}
 
 	return release
-}
-
-func (a *Api) sendPatchNotes() {
-	// get release notes
-	release := getGithubReleaseNotes()
-
-	log.Debug().Any("Release", release).Msg("Github Release")
-
-	releaseBody := strings.ReplaceAll(release.Body, "@ColeHassett", "<@108736074557239296>")
-	releaseBody = strings.ReplaceAll(releaseBody, "@colinthatcher", "<@128629520742744065>")
-
-	embed := &discordgo.MessageEmbed{
-		Title:       fmt.Sprintf("🤯 New PlayDate Release: %s 🤯", release.Name),
-		Description: getFirstNRunes(releaseBody, 4096),
-		Color:       0xfadde6,
-		Timestamp:   release.PublishedAt.Format(time.RFC3339), // Discord expects ISO 8601 for timestamp
-	}
-
-	_, err := a.dg.ChannelMessageSendEmbed(Config.DiscordConfig.ChannelID, embed)
-	if err != nil {
-		log.Err(err).Msg("Failed to send patch notes embed")
-	}
 }
 
 // In case the string is too long for the embed somehow
